DEV: 202220521

+++ ada-boost-implementations/code-python/generation_utils.py
@@ -8,7 +8,7 @@ def make_separable_labels(x_data, rho=1e-1):
     n_samples, n_features = x_data.shape[0], x_data.shape[1]
     buff = np.random.uniform(-1, 1, n_features)
     v_normal = buff/np.linalg.norm(buff)
-    y_data = np.sign(np.squeeze(np.matmul(x_data, v_normal)))
+    y_data = np.sign(np.squeeze(np.matmul(x_data, v_normal))).astype(int)
     x_shift = np.tile(v_normal * rho, (n_samples,1))
     x_shift = np.multiply(x_shift, y_data[:, np.newaxis])
     return x_data + x_shift, y_data, v_normal
